refactor: move onboarding helpers into repository client
Both repository and onboarding endpoints live on the Repository Service,
so colocating them avoids a one-function file and a cross-file helper
export.
